Add mirroring to SingStreamServer and Client SingleStream Server is done

+++ Stream.hpp
@@ -31,7 +31,7 @@ class Stream
   ~Stream();
 
   ssize_t send(const char *buffer, int len);
-  ssize_t receive(char *buffer, int len, int timeout = 0);
+  ssize_t receive(char *buffer, size_t len, int timeout = 0);
 
   string getPeerIP();
   int getPeerPort();
